codex(refactor): extract compose input style helpers
